requirements.txt
tensorflow-gpu==2.4.0
keras_applications
-e setup.py